requirements.txt
qrcode[pil]
pillow